==> token-health/notebooks/preprocessing/1-tvl-chains.ipynb <==
# Preprocessing TVL chains

## Libraries

import pandas as pd
import numpy as np
import json

# Function to load JSON
def loadJSON(filepath):
    with open(filepath) as file:
        return json.load(file)

# Loading the JSON and converting it in a dataframe
jsonFile = loadJSON('../../data/json/all-chains.json')

df = pd.DataFrame(jsonFile)
print(df.shape)
df.head()

## Removing non usable data

# Checking basic info (specially nulls)
df.info()

# Dropping null values and checking again
df = df.dropna(subset=['tokenSymbol'])
df.info()

## Something to point out...
Once we get to the step of getting the information of every token from CoinMarketCap API, we are going to have some issues with "ETHF", "FXTL" and "-" (tokenSymbol property). In order to avoid those API errors, we are removing them

# Initializing the list with those values
values_to_remove = ['ETHF', 'FXTL', '-']

# Filter out rows where tokenSymbol is in values_to_remove
filtered_df = df[~df['tokenSymbol'].isin(values_to_remove)]

filtered_df.to_json('../../data/json/tvl-chains-symbol.json', orient='records', indent=4)
