==> token-health/notebooks/preprocessing/2-tvl-chains-fill-cmcId.ipynb <==
# Filling missing information from the tokens

## Libraries

import pandas as pd
import numpy as np
import os
import json
import requests

## Loading all the chains and its token

# Function to load JSON
def loadJSON(filepath):
    with open(filepath) as file:
        return json.load(file)

df = pd.DataFrame(loadJSON('../../data/json/tvl-chains-symbol.json'))

print(df.shape)
df.head()

## Getting the rows without cmcId

null_cmcId_df = df[df['cmcId'].isnull()]
null_cmcId_df

## Preparing the tokens to be passed to the endpoint

# Setting up the API key from the env
API_KEY = os.getenv('COINMARKETCAP_MARCO_KEY')

# Getting all the symbols
symbols = null_cmcId_df['tokenSymbol'].tolist()

# Convert the list of symbols to a comma-separated string to pass it to the endpoint
symbol_string = ','.join(symbols)

# Display the string of symbols
print("Comma-separated string of symbols:")
print(symbol_string)

## Requesting token information

# URL and headers for the CoinMarketCap API call
url = 'https://pro-api.coinmarketcap.com/v2/cryptocurrency/info'

# API key from environment
API_KEY = os.getenv('COINMARKETCAP_MARCO_KEY')


headers = {
    'Accepts': 'application/json',
    'X-CMC_PRO_API_KEY': API_KEY,
}

# Parameters for the request
parameters = {
    'symbol': symbol_string,
}

# Make the request
response = requests.get(url, headers=headers, params=parameters)

# Check the status of the response
api_response = {}
if response.status_code == 200:
    result = response.json()
    
    # Create a dictionary with the API response and the counter
    api_response = {
        'data': result.get('data', {}),
    }

    # Process and count the information
    for symbol, info in result['data'].items():
        api_response['data'][symbol] = info
    
    print(f"Successful")
else:
    print(f"Error: {response.status_code} - {response.json()}")

## Fill CoinMarketCap ID
As these tokens do not have a specific cmcId, we are going to get several tokens with the same symbol. In order to get a single one, we look for a coincidence from the name from Defillama and the one from CoinMarketCap.

# Storing tokens without coincidence
symbol_without_coincidence = {'data': {}}

for symbol, coins in api_response['data'].items():
    # Identifying the row with the current symbol and its id
    row_with_symbol = null_cmcId_df[null_cmcId_df['tokenSymbol'] == symbol]
    row_with_symbol_id = row_with_symbol.index[0]

    for i, coin in enumerate(coins):
        # Check if the the CoinMarketCap token name includes the one from Defillama
        if row_with_symbol['name'].str.lower().iloc[0] in coin['name'].lower():
            # Replacing the cmcId on the original df and the null one
            df.at[row_with_symbol_id, 'cmcId'] = coin['id']
            null_cmcId_df.at[row_with_symbol_id, 'cmcId'] = coin['id']
            break
        # Add to the dictionary in case there's no coincidence
        if i == len(coins)-1: symbol_without_coincidence['data'][symbol] = coins

# Checking what has changed
null_cmcId_df

After checking the filled ids in Defillama and CoinMarketCap, we remove the ones that don't match and fix some ids that are incorrect

# EMPIRE
df.drop(index=102, inplace=True)
# Mode
df.at[133, 'cmcId'] = '31016'
# Dexit
df.drop(index=151, inplace=True)
# Heiko
df.drop(index=156, inplace=True)
# Parallel
df.drop(index=157, inplace=True)
# Crescent
df.drop(index = 164, inplace = True)
# Tlchain
df.drop(index=181, inplace=True)
# Libre
df.at[185, 'cmcId'] = '27114'
# Bone
df.drop(192, inplace=True)

Now there's less null values in **cmcId!!** :)

# Getting again only the null values
null_cmcId_df = df[df['cmcId'].isnull()]
null_cmcId_df

## Checking and replacing manually
Comparing information from Defillama and CoinMarket to see if there's a coincidence and they have market cap information.

### Zen - Horizen

df.drop(index=56, inplace=True)

### DG - Dogechain

df.at[99, 'cmcId'] = '21414'
df.at[99, 'tokenSymbol'] = "DC"

### WHALE - White Whale

df.at[199, 'cmcId'] = '16121'

### BEAM - Beam

df.at[206, 'cmcId'] = '28298'
df.at[206, 'tokenSymbol'] = "BEAM"

The rest doesn't have enough information or are different from the ones in CoinMarketCap and CoinGecko

# Dropping the rows with null walues in cmcId since we cannot replace them
df = df.dropna(subset=['cmcId'])
df.info()

We finally have all the cmcIds complete!

# Converting all to strings
df['cmcId'] = df['cmcId'].apply(lambda x: str(x))

## Checking duplicated values

duplicates = df[df.duplicated(subset=['cmcId'], keep=False)]
duplicates.sort_values(by='cmcId')

# There's information for Bitcoincash but not for smartBCH in CoinMarketCap
df.drop(index=101, inplace=True)

# The information of BSC is not in CoinMarketCap, but the market cap of Binance is more than 80 000 000 000, so we can get rid of it
df.drop(index=5, inplace=True)
df.drop(index=33, inplace=True)

# Incorrect id
df.at[51, 'cmcId'] = '7950'

## Save to JSON

df.to_json('../../data/json/main-defi-tokens.json', orient='records', indent=4)
