==> token-health/notebooks/preprocessing/3-tvl-chains-fill-gecko-id.ipynb <==
# Filling missing information from the tokens

## Libraries

import pandas as pd
import numpy as np
import os
import json
import requests

## Loading all the chains and its token

# Function to load JSON
def loadJSON(filepath):
    with open(filepath) as file:
        return json.load(file)

df = pd.DataFrame(loadJSON('../../data/json/main-defi-tokens.json'))

print(df.shape)
df.head()

## Getting the rows without gecko_id

null_gecko_df = df[df['gecko_id'].isnull()]
null_gecko_df

Since it's only 4 rows missing, we are completing the null values manually through the CoinGecko web platform

### RON - Ronin

df.at[43, 'gecko_id'] = 'ronin'

### QGOV - Q Protocol

df.at[68, 'gecko_id'] = 'q-protocol'

### ZENIQ - Zeniq

df.at[170, 'gecko_id'] = 'zeniq'

### BEAM - Beam

df.at[193, 'gecko_id'] = 'beam-2'

# Checking basic info to see how the null situation is at the moment
df.info()

df.to_json('../../data/json/main-tokens-clean.json', orient='records', indent=4)
