==> token-health/notebooks/preprocessing/7-token-category.ipynb <==
**Libraries**

import json
import pandas as pd

**Loading crypto_final.csv and tokens-tvl-full.json files**

with open('../../data/json/tokens-tvl-full-close-cap.json', 'r') as f:
    crypto_data = json.load(f)


csv_data = pd.read_csv('../../../logical-model/data/csv/crypto_final.csv')

**Creating a dictionary to map coin IDs to their categories from the CSV**

category_map = dict(zip(csv_data['id'].astype(str).str.strip(), csv_data['category']))

tokens_full_category = {"data": {}}

**Iterating and replacing categories**

for coin_info in crypto_data['data'].values():
    coin_id = str(coin_info['id'])  # Ensure the ID is a string
    category_num = -1
    if coin_id in category_map:
        if(category_map[coin_id] == 'AI'):
            category_num = 0
        elif(category_map[coin_id] == 'Gaming'):
            category_num = 1
        elif(category_map[coin_id] == 'RWA'):
            category_num = 2
        elif(category_map[coin_id] == 'Meme'):
            category_num = 3
        else:
            category_num = 4
        
        coin_info['category'] = category_num  # Replace with the category from the CSV
    
    # Add the modified coin to the new JSON
    tokens_full_category["data"][coin_id] = coin_info

with open('../../data/json/tokens-tvl-full-close-cap-category.json', 'w') as f:
    json.dump(tokens_full_category, f, indent=4)
