==> token-health/notebooks/preprocessing/9-correlation.ipynb <==
# Creating the correlation field

## Libraries

import pandas as pd
import matplotlib.pyplot as plt
import json
import numpy as np
pd.set_option('display.max_rows', None)

# Function to load JSON
def loadJSON(filepath):
    with open(filepath) as file:
        return json.load(file)

df = pd.DataFrame(loadJSON('../../data/json/tokens-pre-final.json'))

print(df.shape)
df.head()

## Calculation correlation between pairs

# Convert Date to datetime
df['Date'] = pd.to_datetime(df['Date'], format='%d/%m/%Y')

# Set Date as index (for rolling operations)
df.set_index('Date', inplace=True)

# Group by 'Token' and calculate rolling correlation for each pair of columns
def calculate_correlation(group):
    correlation_results = {}
    correlation_results['TVL_Price'] = group['TVL'].rolling(window=15).corr(group['Price'])
    correlation_results['TVL_Market_Cap'] = group['TVL'].rolling(window=15).corr(group['Market cap'])
    correlation_results['Price_Market_Cap'] = group['Price'].rolling(window=15).corr(group['Market cap'])
    
    return pd.DataFrame(correlation_results)

# Apply the function to each group
correlation_df = df.groupby('Token').apply(calculate_correlation).reset_index(level=0, drop=True)

# Combine the correlation results with the original DataFrame
df = pd.concat([df, correlation_df], axis=1)

# Fill NaN values with 0 (or handle as needed)
df.fillna(0, inplace=True)
df.replace([np.inf, -np.inf], 0, inplace=True)

# Reset the index to make Date a regular column again
df.reset_index(inplace=True)

# Checking the generated values
df.head(50)

## Calculating the average correlation

# Calculating the average correlation
df['correlation_average'] = df[['TVL_Price', 'TVL_Market_Cap', 'Price_Market_Cap']].mean(axis=1)

df.head(20)

# Checking basic stats
df['correlation_average'].describe()

## Generating the Correlation field

# Defining the threshold
threshold = 0.8
# Generating the column
df['Correlation'] = (df['correlation_average'] >= threshold).astype(int)

df.head(20)

# Checking the amount of values
df['Correlation'].value_counts()

# Just selecting the important features
final_df = df[['Date', 'Token', 'Name', 'TVL', 'Market cap', 'Price', 'Correlation', 'Class']]
final_df.head()

# Converting to date
final_df['Date'] = pd.to_datetime(final_df['Date']).dt.strftime('%d-%m-%Y')

final_df.head()

# Saving in a JSON file
final_df.to_json('../../data/json/tokens-final.json', orient='records', indent=4)
